[Chore] LocationFilterVC View 분리 (#143)
